+++ UniTAS/Plugin/Implementations/MainThreadSpeedController.cs
@@ -30,7 +30,7 @@ public class MainThreadSpeedControl : IMainThreadSpeedControl, IOnUpdate, IOnMov
     private float _lastTime;
     private float _speedMultiplier = 1f;
 
-    private float _remainingTime;
+    private double _remainingTime;
 
     public MainThreadSpeedControl(IPatchReverseInvoker patchReverseInvoker, ITimeEnv timeEnv)
     {
@@ -54,7 +54,7 @@ public class MainThreadSpeedControl : IMainThreadSpeedControl, IOnUpdate, IOnMov
             return;
         }
 
-        var waitMilliseconds = waitTime * 1000f;
+        var waitMilliseconds = waitTime * 1000.0;
         var waitMillisecondsInt = (int)waitMilliseconds;
 
         if (waitMillisecondsInt > 0)
@@ -63,7 +63,7 @@ public class MainThreadSpeedControl : IMainThreadSpeedControl, IOnUpdate, IOnMov
         }
 
         // lost time is added to the remaining time
-        _remainingTime = (waitMilliseconds - waitMillisecondsInt) / 1000f;
+        _remainingTime = (waitMilliseconds - waitMillisecondsInt) / 1000.0;
     }
 
     public void OnMovieRunningStatusChange(bool running)
